+++ themes/grav/app/forms/fields/collections.js
@@ -177,10 +177,19 @@ export default class CollectionsField {
                 item.find('[' + prop + '], [_' + prop + ']').each(function() {
                     let element = $(this);
                     let indexes = [];
+                    let array_index = null;
                     let regexps = [
                         new RegExp('\\[(\\d+|\\*|' + currentKey + ')\\]', 'g'),
                         new RegExp('\\.(\\d+|\\*|' + currentKey + ')\\.', 'g')
                     ];
+
+                    // special case to preserve array field index keys
+                    if (prop === 'name' && element.data('gravArrayType')) {
+                        const match_index = element.attr(prop).match(/\[\d+\]$/);
+                        if (match_index) {
+                            array_index = match_index[0];
+                        }
+                    }
 
                     if (hasCustomKey && !observedValue) {
                         element.attr(`_${prop}`, element.attr(prop));
@@ -198,8 +207,11 @@ export default class CollectionsField {
 
                     let matchedKey = currentKey;
                     let replaced = element.attr(prop).replace(regexps[0], (/* str, p1, offset */) => {
+                        let extras = '';
+                        if (array_index) { extras = array_index; }
+
                         matchedKey = indexes.shift() || matchedKey;
-                        return `[${matchedKey}]`;
+                        return `[${matchedKey}]${extras}`;
                     });
 
                     replaced = replaced.replace(regexps[1], (/* str, p1, offset */) => {