+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Fixed 2FA regenerate for Flex Users
     * Added missing closing </li> in language loops
     * Fixed issue with nested `list` fields both utilizing the custom `key` functionality
+    * Fixed issue with `array` field nested in `list` that were losing their index order when the list reordered
 
 # v1.9.7
 ## 06/21/2019